libnice-crash: Add reproduction test based on kurento-hello-world
Only difference is that all candidates are forced to be only TCP, and
that the WebRtcEndpoint and the Pipeline objects are hard-deleted in
Java, instead of following a proper release procedure.

This triggers something in the way the inner objects are released via
the Client API, and that causes libnice to crash.